arachnida/requirements.txt
beautifulsoup4
pyexiftool
pygame
Pillow